+++ src/Symfony/Component/Messenger/Middleware/ChainMiddleware.php
@@ -17,6 +17,7 @@ use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Stamp\BusNameStamp;
 use Symfony\Component\Messenger\Stamp\ChainStamp;
 use Symfony\Component\Messenger\Stamp\DispatchAfterCurrentBusStamp;
+use Symfony\Component\Messenger\Stamp\DispatchOnFailureStamp;
 use Symfony\Component\Messenger\Stamp\NoAutoAckStamp;
 use Symfony\Component\Messenger\Stamp\ReceivedStamp;
 use Symfony\Component\Messenger\Stamp\SentStamp;
@@ -79,6 +80,10 @@ final class ChainMiddleware implements MiddlewareInterface
 
         if (null === $next->last(BusNameStamp::class) && null !== $busNameStamp = $envelope->last(BusNameStamp::class)) {
             $next = $next->with($busNameStamp);
+        }
+
+        if (null === $next->last(DispatchOnFailureStamp::class) && null !== $failureStamp = $envelope->last(DispatchOnFailureStamp::class)) {
+            $next = $next->with($failureStamp);
         }
 
         $this->bus->dispatch($next);