+++ src/Symfony/Component/Messenger/Resources/config/messenger.php
@@ -20,6 +20,7 @@ use Symfony\Component\Messenger\Bridge\Beanstalkd\Transport\BeanstalkdTransportF
 use Symfony\Component\Messenger\Bridge\MongoDb\Transport\MongoDbTransportFactory;
 use Symfony\Component\Messenger\Bridge\Redis\Transport\RedisTransportFactory;
 use Symfony\Component\Messenger\EventListener\AddErrorDetailsStampListener;
+use Symfony\Component\Messenger\EventListener\DispatchOnFailureListener;
 use Symfony\Component\Messenger\EventListener\DispatchPcntlSignalListener;
 use Symfony\Component\Messenger\EventListener\ReleaseDeduplicationLockOnFailureListener;
 use Symfony\Component\Messenger\EventListener\ResetMemoryUsageListener;
@@ -35,6 +36,7 @@ use Symfony\Component\Messenger\Middleware\ChainMiddleware;
 use Symfony\Component\Messenger\Middleware\DecodeFailedMessageMiddleware;
 use Symfony\Component\Messenger\Middleware\DeduplicateMiddleware;
 use Symfony\Component\Messenger\Middleware\DispatchAfterCurrentBusMiddleware;
+use Symfony\Component\Messenger\Middleware\DispatchOnFailureMiddleware;
 use Symfony\Component\Messenger\Middleware\FailedMessageProcessingMiddleware;
 use Symfony\Component\Messenger\Middleware\HandleMessageMiddleware;
 use Symfony\Component\Messenger\Middleware\LoggingMiddleware;
@@ -136,6 +138,13 @@ return static function (ContainerConfigurator $container) {
             ->args([
                 service('messenger.routable_message_bus'),
             ])
+
+        ->set('messenger.middleware.dispatch_on_failure', DispatchOnFailureMiddleware::class)
+            ->args([
+                service('messenger.routable_message_bus'),
+                service('logger')->ignoreOnInvalid(),
+            ])
+            ->tag('monolog.logger', ['channel' => 'messenger'])
 
         ->set('messenger.middleware.validation', ValidationMiddleware::class)
             ->args([
@@ -264,6 +273,14 @@ return static function (ContainerConfigurator $container) {
                 abstract_arg('failure transports'),
                 service('logger')->ignoreOnInvalid(),
                 abstract_arg('failure transports by name'),
+            ])
+            ->tag('kernel.event_subscriber')
+            ->tag('monolog.logger', ['channel' => 'messenger'])
+
+        ->set('messenger.failure.dispatch_on_failure_listener', DispatchOnFailureListener::class)
+            ->args([
+                service('messenger.routable_message_bus'),
+                service('logger')->ignoreOnInvalid(),
             ])
             ->tag('kernel.event_subscriber')
             ->tag('monolog.logger', ['channel' => 'messenger'])