+++ src/Symfony/Component/Messenger/Tests/Middleware/ChainMiddlewareTest.php
@@ -23,6 +23,7 @@ use Symfony\Component\Messenger\Stamp\BusNameStamp;
 use Symfony\Component\Messenger\Stamp\ChainStamp;
 use Symfony\Component\Messenger\Stamp\DelayStamp;
 use Symfony\Component\Messenger\Stamp\DispatchAfterCurrentBusStamp;
+use Symfony\Component\Messenger\Stamp\DispatchOnFailureStamp;
 use Symfony\Component\Messenger\Stamp\NoAutoAckStamp;
 use Symfony\Component\Messenger\Stamp\ReceivedStamp;
 use Symfony\Component\Messenger\Stamp\SentStamp;
@@ -193,6 +194,48 @@ class ChainMiddlewareTest extends MiddlewareTestCase
         $this->expectExceptionMessage('A message handled by the batch handler "Closure" cannot carry a "Symfony\\Component\\Messenger\\Stamp\\ChainStamp".');
 
         $middleware->handle($envelope, $this->getStackAdding(new NoAutoAckStamp(new HandlerDescriptor(static function () {}))));
+    }
+
+    public function testDispatchOnFailureStampIsCopiedToTheNextMessage()
+    {
+        $second = new SecondMessage();
+        $failureStamp = new DispatchOnFailureStamp(new ThirdMessage());
+        $envelope = new Envelope(new DummyMessage('first'), [$failureStamp, new ChainStamp($second)]);
+
+        $bus = $this->createMock(MessageBusInterface::class);
+        $bus->expects($this->once())
+            ->method('dispatch')
+            ->with($this->callback(function (Envelope $next) use ($second, $failureStamp) {
+                $this->assertSame($second, $next->getMessage());
+                $this->assertSame([$failureStamp], $next->all(DispatchOnFailureStamp::class));
+
+                return true;
+            }))
+            ->willReturnArgument(0);
+
+        $middleware = new ChainMiddleware($bus);
+        $middleware->handle($envelope, $this->getStackMock());
+    }
+
+    public function testEnvelopeInChainKeepsItsOwnDispatchOnFailureStamp()
+    {
+        $second = new SecondMessage();
+        $ownFailureStamp = new DispatchOnFailureStamp(new ThirdMessage());
+        $envelope = new Envelope(new DummyMessage('first'), [new DispatchOnFailureStamp(new DummyMessage('failure')), new ChainStamp(new Envelope($second, [$ownFailureStamp]))]);
+
+        $bus = $this->createMock(MessageBusInterface::class);
+        $bus->expects($this->once())
+            ->method('dispatch')
+            ->with($this->callback(function (Envelope $next) use ($second, $ownFailureStamp) {
+                $this->assertSame($second, $next->getMessage());
+                $this->assertSame([$ownFailureStamp], $next->all(DispatchOnFailureStamp::class));
+
+                return true;
+            }))
+            ->willReturnArgument(0);
+
+        $middleware = new ChainMiddleware($bus);
+        $middleware->handle($envelope, $this->getStackMock());
     }
 
     public function testNothingIsDispatchedWhenHandlingFails()